fix(cursorValence): extract shared computeCursorValence; fix Pass mapping to ~0 in ValenceBeatPad
